NIFI-11536: replaced KeyStoreScannerTest and TrustStoreScannerTest with StoreScannerTest
