+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-jetty/src/test/java/org/apache/nifi/web/server/util/StoreScannerTest.java
@@ -1,0 +1,136 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.nifi.web.server.util;
+
+import org.apache.nifi.security.util.TemporaryKeyStoreBuilder;
+import org.apache.nifi.security.util.TlsConfiguration;
+import org.eclipse.jetty.util.resource.Resource;
+import org.eclipse.jetty.util.ssl.SslContextFactory;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.ArgumentCaptor;
+import org.mockito.ArgumentMatchers;
+import org.mockito.Mockito;
+
+import javax.net.ssl.SSLContext;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Paths;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Consumer;
+
+public class StoreScannerTest {
+    private SslContextFactory sslContextFactory;
+    private static TlsConfiguration tlsConfiguration;
+    private static File keyStoreFile;
+    private static File trustStoreFile;
+    private static Map<File, StoreScanner> filesToScannerMap;
+
+    @BeforeAll
+    public static void initClass() {
+        tlsConfiguration = new TemporaryKeyStoreBuilder().build();
+        keyStoreFile = Paths.get(tlsConfiguration.getKeystorePath()).toFile();
+        trustStoreFile = Paths.get(tlsConfiguration.getTruststorePath()).toFile();
+    }
+
+    @BeforeEach
+    public void init() throws IOException {
+        sslContextFactory = Mockito.mock(SslContextFactory.class);
+        Resource keyStoreResource = Mockito.mock(Resource.class);
+        Mockito.when(keyStoreResource.getFile()).thenReturn(keyStoreFile);
+        Mockito.when(sslContextFactory.getKeyStoreResource()).thenReturn(keyStoreResource);
+
+        Resource trustStoreResource = Mockito.mock(Resource.class);
+        Mockito.when(trustStoreResource.getFile()).thenReturn(trustStoreFile);
+        Mockito.when(sslContextFactory.getTrustStoreResource()).thenReturn(trustStoreResource);
+
+        final StoreScanner keyStoreScanner = new StoreScanner(sslContextFactory, tlsConfiguration, sslContextFactory.getKeyStoreResource());
+        final StoreScanner trustStoreScanner = new StoreScanner(sslContextFactory, tlsConfiguration, sslContextFactory.getTrustStoreResource());
+        filesToScannerMap = new HashMap<>();
+        filesToScannerMap.put(keyStoreFile, keyStoreScanner);
+        filesToScannerMap.put(trustStoreFile, trustStoreScanner);
+    }
+
+    @Test
+    public void testFileAdded() throws Exception {
+        for (final Map.Entry<File, StoreScanner> entry : filesToScannerMap.entrySet()) {
+            final File file = entry.getKey();
+            final StoreScanner scanner = entry.getValue();
+            scanner.fileAdded(file.getAbsolutePath());
+
+            final ArgumentCaptor<Consumer> consumerArgumentCaptor = ArgumentCaptor.forClass(Consumer.class);
+            Mockito.verify(sslContextFactory).reload(consumerArgumentCaptor.capture());
+
+            final Consumer<SslContextFactory> consumer = consumerArgumentCaptor.getValue();
+            consumer.accept(sslContextFactory);
+            Mockito.verify(sslContextFactory).setSslContext(ArgumentMatchers.any(SSLContext.class));
+            Mockito.clearInvocations(sslContextFactory);
+        }
+    }
+
+    @Test
+    public void testFileChanged() throws Exception {
+        for (final Map.Entry<File, StoreScanner> entry : filesToScannerMap.entrySet()) {
+            final File file = entry.getKey();
+            final StoreScanner scanner = entry.getValue();
+            scanner.fileChanged(file.getAbsolutePath());
+
+            final ArgumentCaptor<Consumer> consumerArgumentCaptor = ArgumentCaptor.forClass(Consumer.class);
+            Mockito.verify(sslContextFactory).reload(consumerArgumentCaptor.capture());
+
+            final Consumer<SslContextFactory> consumer = consumerArgumentCaptor.getValue();
+            consumer.accept(sslContextFactory);
+            Mockito.verify(sslContextFactory).setSslContext(ArgumentMatchers.any(SSLContext.class));
+            Mockito.clearInvocations(sslContextFactory);
+        }
+    }
+
+    @Test
+    public void testFileRemoved() throws Exception {
+        for (final Map.Entry<File, StoreScanner> entry : filesToScannerMap.entrySet()) {
+            final File file = entry.getKey();
+            final StoreScanner scanner = entry.getValue();
+            scanner.fileRemoved(file.getAbsolutePath());
+
+            final ArgumentCaptor<Consumer> consumerArgumentCaptor = ArgumentCaptor.forClass(Consumer.class);
+            Mockito.verify(sslContextFactory).reload(consumerArgumentCaptor.capture());
+
+            final Consumer<SslContextFactory> consumer = consumerArgumentCaptor.getValue();
+            consumer.accept(sslContextFactory);
+            Mockito.verify(sslContextFactory).setSslContext(ArgumentMatchers.any(SSLContext.class));
+            Mockito.clearInvocations(sslContextFactory);
+        }
+    }
+
+    @Test
+    public void testReload() throws Exception {
+        for (final Map.Entry<File, StoreScanner> entry : filesToScannerMap.entrySet()) {
+            final StoreScanner scanner = entry.getValue();
+            scanner.reload();
+
+            final ArgumentCaptor<Consumer> consumerArgumentCaptor = ArgumentCaptor.forClass(Consumer.class);
+            Mockito.verify(sslContextFactory).reload(consumerArgumentCaptor.capture());
+
+            final Consumer<SslContextFactory> consumer = consumerArgumentCaptor.getValue();
+            consumer.accept(sslContextFactory);
+            Mockito.verify(sslContextFactory).setSslContext(ArgumentMatchers.any(SSLContext.class));
+            Mockito.clearInvocations(sslContextFactory);
+        }
+    }
+}
